backend/requirements.txt
flask
flask-cors==3.0.6
flask-mongoengine==0.9.5
flask_script
uwsgi
pyjwt
xlrd
diff-match-patch